Add `Configuration` entity
